🐛 Don’t edit the original graph during tracing
The way Graph objects work, any changes made to the graph are written back to the file. This means that the copy of the graph *must* be used during tracing, otherwise the original input graph is edited (and closing the editor asks if you want to save changes). We now get the absolute path of the graph and program files, update those paths to add “_tracing” at the end, and give those to TraceRunner instead.