✨ Implement entering & exiting contexts backwards
When going backwards, we need to enter a context whenever one was exited, and vice versa. This means that we can’t just have an END_CONTEXT step type, because we need to know what context we are entering. So instead, add TraceStep::endOfContext, and record the step type properly for end of contexts.