🐛 Fix conditions for match and apply
Because we now have a TraceStep for the </match> element, we can’t use the context stack to determine if rule matching is available. It also means that we cannot use it to determine if rule application is available, because we will be in a <rule> context twice for each rule.