🐛 Don’t exit failed rule matches twice
When a failed rule match was found, we skipped the loop which parsed the </match> token, but we still manually added the TraceStep with type RULE_MATCH_FAILED and endOfContext = true. This meant that when stepping over a failed rule match, the context would be exited twice, causing an empty stack error.