✨ Enter procedures in reverse when stepping back
In order to trace backwards through a procedure, we need to enter it from the end, not the beginning. In this case, we let the search loop around until it finds the next procedure delcaration, which means that searching for the next token is starting from the end of the procedure, as required.