🐛 Don’t restore snapshots in rulesets
If a rule fails in a ruleset, we don’t want to restore the loop iteration snapshot unless it’s the last rule in the set, because a future rule may still succeed. If it’s the last one in the set, then that’s treated as a failure, so the snapshot should be restored.